requirements.txt
pydub
whisper @ git+https://github.com/openai/whisper.git
typer[all]
pyyaml